Fix atomic_compare_exchange_strong() function usage.
In some of the SYCL CTS atomic tests the
atomic_compare_exchange_strong() API is used incorrectly where the
second parameter (expected value) is passed as a pointer instead of by
reference.  This is most likely a mixup between SYCL 1.2 and 1.2.1 where
the API changed from using a pointer to a reference.