Changed sub buffer construction parameters
According to the SYCL specification clarifications from
https://github.com/KhronosGroup/SYCL-Docs/pull/39, while constructing
sub buffer the offset and range specified by baseIndex and subRange
which passed to `buffer(buffer)` constructor together must represent a
contiguous region of the original SYCL buffer. Besides, the sum of
`baseIndex` and `subRange` in any dimension must not exceed the parent
buffer (`b`) size (`bufferRange`) in that dimension.

This patch changes an offset for sub buffer so it forms a contiguous region,
and calculates sub buffer range based on given offset and parent buffer range.

Patch by Ivan Karachun.

Signed-off-by: Alexey Bader <alexey.bader@intel.com>