Change accessor image channel data types when using DPCPP CUDA backend
The DPCPP CUDA backend currently only support reading from and writing
to image accessors with data type corresponding to the channel data type
of the underlying image.

This changes the channel types of images used in the accessor tests to
correspond to the supported accessor data types when using the DPCPP
CUDA backend.

Signed-off-by: Steffen Larsen <steffen.larsen@codeplay.com>