Test that handler invoke APIs pass correct indices to kernels
This adjusts the handler invoke API tests to not only check if a kernel
is launched at all, but also whether it receives the correct indices
spanning the requested iteration space. In particular, each kernel
thread now writes its corresponding global work item id into a buffer
that is then checked on the host.

Some test cases now receive a non-zero offset, and the output buffer is
padded with canary values to detect whether an implementation launches
threads outside the requested range.