Add support for strided/transposed copies
This adds support for copying (and other explicit memory operations) of
strided and transposed buffers. Strided buffers are buffers that are
accessed with a ranged accessor, while transposed copies switch up the
dimensions of a target buffer. For example, instead of copying from
[4,8] to [4,8], a transposed copy copies from [4,8] to [8,4].

This greatly increases the number of test case permutations, potentially
surfacing additional issues within SYCL implementations.