Add extensive mode cmake option
Improving data type coverage up to the full one can give an
unreasonable performance overhead for the specific functionality,
like accessors.
Type coverage for the default mode should be balanced in terms of
coverage and performance, but there should be an option to make all
extra checks when performance doesn't matter.