Minimize data race possibility
As mentioned in PR #124 there is a possibility of data race on
writing even if values are expected to be bit-wise equal and data
race is UB in C++. So we are required to collect any information
on per-work-item basis to avoid simulationeous write access.

Also we should provide a sane message on test failure. Thus the test
should actually init any variables with the default values and change
them later. It makes impossible to bit-and any boolean flag.
Therefore we are required to split data flow further and provide
separate storage space for every logical item. Also it requires us to
remove separate submit calls, so that every value is written once.

To do so we need to have unique offset for every logical work item.
It forces us to use linear logical local id and group id. Both these
calls can possibly return invalid values and we need to handle these
cases somehow.
Simple overflow check with the default value trick mentioned earlier
makes it possible to detect these failures. Code behavior is well
defined for the case of SYCL implementation errors leading to the
too big offset values. Still any SYCL implementation providing
duplicate id values will result in data race accessing the same
offset. Because that's an UB we can not really expect test even to
output anything in such case, still if by any means only data
corruption occurs, test output would still be helpfull by showing up
values missing and broken.

There are also some features that can be used to avoid UB in the test
code entirely:
  - atomic access (using unsigned int for portability and extra
    checks to ensure id values are not truncated upon conversion)
  - reduction variables with sycl::bit_and
We are not using them to minimize dependency on non-related features.
We still depend on linear id retrieval for group and h_item though.