Build supported test categories during CI
Build the set of supported test categories for each SYCL implementation
during CI runs. This uses the generated exclude filters within the ./ci
directory to limit the set of tests to compile for each implementation.

Limit parallel jobs to 2 (-j3 is Ninja's default on GitHub hosted
runners), as we otherwise risk running out of memory.