Fix out-of-scope usage of `IMPL_VERSION` in Dockerfiles
Docker build args are scoped to a particular build stage (`FROM`).
By starting a new one, previous args go out of scope. Ensure that
`IMPL_VERSION` is available inside the actual build by adding an
additional `ARG` directive, as well as a check to confirm that the
argument is set.