Overhaul device selection mechanism
This changes the way of selecting the platform/device to run the CTS on
from a hard-coded list of options to a mechanism based on substring
matching.

The options `--platform`, `-p`, `--device` and `-d` are replaced by a
single `--device` option that receives a string that is interpreted as a
ECMAScript regular expression for matching against all available devices
reported by the SYCL implementation. Additionally, a new
`--list-devices` option is introduced that lists all aforementioned
devices, additionally indicating the device currently matched against
by `--device`.

Additional changes:
   - If no `--device` is given, fall back to the default selector.
   - Replace `[host|opencl]_[platform|device]_name` CMake configuration
     parameters by `SYCL_CTS_CTEST_DEVICE`.
   - Update runtests.py to receive single `--device` option.
   - Fix some additional issues in runtests.py script.