Remove /subsystem:console from LINK_FLAGS for DPC++ on Windows
The clang-cl compiler doesn't support the '/subsystem:console' link flag.
CMake sets up the flag in the 'CMAKE_CXX_CREATE_CONSOLE_EXE' variable,
not in CMAKE_CREATE_CONSOLE_EXE one (I'm not sure in which CMake version
the variable was changed.) The patch replaces the flag in the
CMAKE_CXX_CREATE_CONSOLE_EXE with an empty string.