Avoid tabs in command lines
This commit updates cmake/FindComputeCpp.cmake to avoid the use of tabs
in command lines. The tabs are valid but cause issues when copying and
pasting commands to run them standalone, as even the pasting of a tab
may trigger tab completion, depending on the shell and shell settings.

The tabs were a workaround for older CMake versions, but a sufficiently
recent CMake is already required.