Encode SYCL impl version in CI container image tag
In order to be able to bump a container's version as part of a PR but
without disrupting CI on other branches, we have to include the version
in the container image tag.

A downside that comes with this is that we now need to manually keep the
versions in sync between the two workflows.

Since the ComputeCpp DockerHub repository has not been updated to use
versioned tags, we don't specify a version for ComputeCpp for the time
being.