Fix accessor_api_image_core test for image arrays
When the acc_coord_tag::use_normalized_upper tag is used to compute
the expected values in the linear filtering mode, the value for idx + 1
is used as expected (see the return get_expected_value_linear(idx + 1)
line in the body of the T get_expected_value(image_id_t<dim, target> idx)
method of the image_accessor_api_sampled_r class). This fact is taken into
account in the image_access<dims>::get_normalized() function and works
fine for a single image but not for image arrays since for the image
array, a correct slice should be the next one when the linear filtering
mode is used and the test is run for acc_coord_tag::use_normalized_upper
tag.

Signed-off-by: Pavel Samolysov <pavel.samolysov@intel.com>