Do not increment the last component of idx for image_array
We should increment all the components of idx but the last one if we are
working with target::image_array in the get_expected_value method for
the linear filtration and coordTag equals to
acc_coord_tag::use_normalized_upper. The last component is not a
coordinate, it is the index of an image in the array to read from.

Signed-off-by: Pavel Samolysov <pavel.samolysov@intel.com>