Require size member function for assign and are_equal
Overloads of value_operations::assign and value_operations::are_equal
for types with subscript operator uses the size member function for
iterating the members. However it only checks for the subscript
operator. This can cause compile-time errors when using these with
types such as user_struct that define the subscript operator and a
comparison operator but not the size member function.

This commit further restricts the aforementioned overloads by requring
the size member function. This is done through adding a new type trait
for checking a type for the size member function, namely has_size, and
a shortcut trait has_subscript_and_size for the conjunction of
has_subscript_operator and has_size.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>