Merge developer docs into a single file
This is to avoid having to replicate (manually, or with a script) the
table of contents from the individual chapters in the main entry file.
AsciiDoctor does not support including table of contents from other
files (except when including the whole file, which in turn is not yet
supported by GitHub's renderer).