Remove use of is_host and fix default device constructor check
SYCL 2020 removes the is_host member function from SYCL classes in
tandem with removing the requirement that the host device must be
available. This commit removes the uses of the is_host member function
and fixes the expected behavior of the default device constructor to
expect the resulting device to be the same as one constructed with
default_selector_v.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>