Fix strict aliasing violation in check_as_result
This function writes values into a temporary array using one type and
then reads them back from the same array using a different type to
compare against the expected result. This is undefined behavior in C and
C++, so this has been causing a hang when running these tests. The fix
is to do this properly by using separate arrays for the
differently-typed values and a memcpy to copy the bits from the input
type to the output type.