Configure OpenCL proxy before configuring SYCL executable
This commit moves the OpenCL proxy configuration into a separate cmake
file (cmake/AddOpenCLProxy.cmake) and includes it before including
AddSYCLExecutable. This is motivated by CMake on Windows failing to
configure for DPC++ due to the OpenCL proxy picking Clang as a C
compiler due to FindDPCPP overwriting the compilers.

Additionally, since the OpenCL proxy needs options to be correctly
configured the CTS options are declared before including both the OpenCL
proxy and the SYCL executable.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>