Removed failing deprecated image target accessor tests
check_image_accessor_api_type no longer had a function due to it not
being SYCL 2020 compliant as the image target has been removed. This
commit removes it and things that depended on it.

The suite now fails on the "no member named 'error_category_for' in
namespace 'sycl'" error that is being addressed in other work.