Fix reference builtins for mixing vectors and scalars
The min, max, and clamp reference builtins currently assume that the
type inside the vector argument and the other argument are the same.
However, this is not necessarily the case when using the char vector
alias and the char scalar type as the former uses `int8_t` which may
be defined as `signed` or `unsigned` which are considered different
types from `char` without specified signedness.
This commit relaxes the reference implementations to allow different
scalar type to the element type in the passed vector, requiring it is
implicitly convertible to the element type.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>