Fix array initialization in multi_ptr for no default ctor
The multi_ptr tests may run with no_def_cnstr but will fail to compile
due to the existance of uninitialized arrays in certain tests
requiring the default constructor. This commit adds an initilizer for
these arrays, skipping initialization for def_cnstr as it has no
constructor accepting an integer.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>