Test only remaining legacy get_pointer variants
The accessor legacy tests check for the SYCL 1.2.1 `get_pointer()`
returning a `multi_ptr`, but in SYCL 2020 the only variants remaining of
these are for the deprecated targets `constant_buffer` and `local`. This
commit removes the checks for all non-deprecated accessor targets and
adjusts the `host_buffer` case to conform to the new SYCL 2020 read-only
case.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>