Make accessor test cases templated on non-type arguments
Currently the tests for accessors iterate through all combinations of
valid arguments each time they try to run a specific section for a
specific combination. Though the irrelevant combinations for a given
section run no checks or do any significant work, they still introduce
an overhead that accumulates, causing it to increase drastically when
any of the argument spaces increase, e.g. as the type arguments do in
full-conformance mode.

To decrease the overhead, this commit changes the accessor tests to
use "templated" test cases which each get a combination of arguments
and will not check for their relevant sections under other combinations.
This should not affect the coverage of the tests, but should decrease
the width of the search each section run has to do.

Note that this only currently templates the test cases on access mode,
dimensions, and targets. Including the types in this to further focus
testing runs is left for future work.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>