Fix sub_group_mask_extract_bits expectation and comparison
This commit fixes an issue where the sub_group_mask_extract_bits would
not use a valid comparison operation for marray argument types. This
also revealed a faulty expectation in the test for marray as only the
first element would previously have been checked. This faulty
expectation is also fixed by this patch.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>