Store scan results by group local ID instead of global ID
Since sub-group IDs are not required to follow the same ordering as
the group, the scan test also needs to consider this when computing
the results of the sub-group scan. This commit fixes it so the test
makes no ordering assumption about neither group nor sub-group, avoiding
both faulty expectations and invalid accesses.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>