Fix sub-group 'of' tests
As discovered/discussed in
https://github.com/KhronosGroup/SYCL-CTS/pull/795, the
`predicate_function_of_sub_group` test was using the global linear ID to
perform sub-group 'of' checks.

The checks were only being run on the first sub-group (with ID 0). This
was implicitly relying on the first sub-group containing the work-items
with IDs from 0 up to the sub-group size, as the checks included
conditions like exactly one ID being 1. This is not correct in general,
as the mapping between work-items and sub-group items is not defined.

In addition, this commit also enhances the sub-group tests to check that
*all* sub-groups in the work-group return the correct result. Before we
were either checking only the *first* sub-group
(`predicate_function_of_sub_group`) or checking all sub-groups but
overwriting the results each time, so in effect only checking the *last*
sub-group.

Now the results are initialized to 'true' and each sub-group is run and
its results are merged with all of the other results.