Add test_conformance target with all tests except test_all
test_all is a special test executable including the tests from all other
test executables. For the sake of running the conformance script, it is
both a redundant and a resource heavy task to include test_all as part
of the run. To avoid excessive work, this commit adds a new target
"test_conformance" that includes all test executables except test_all.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>