Improve support detection for USM copy tests
Different USM allocation types are locked behind different aspects.
The USM tests account for this by skipping the test if the aspect
for the destination allocation is not present. However, two tests
(copy and memcpy) have an allocation for the source memory, which
is not checked.

This change adds a "supports_device" method to the USM tests that
takes a queue and returns whether the test supports that queue.
For the copy and memcpy tests, this checks that the source
allocation is supported by the device. All other tests just return
true.