Do not check for undefined behavior in abs.
Unlike OpenCL abs, but like C and C++ abs, in SYCL 2020, taking the
absolute value of a signed type produces a signed result and the
behavior is undefined if the absolute value is not representable (i.e.
if the input is the greatest negative value), so do not check that the
SYCL results match the reference results for this case.