Fix sub-group broadcast test
The test made incorrect assumptions about work-items mapping to
sub-groups, i.e. it expected that sub-groups are formed from subsequent
work-items and thus the last element in the first sub-group of an
nd-range would have its global linear id equal to sub-group size.

However, SYCL specification does not mandate any specific mapping of
work-items into sub-groups and therefore assumptions made by test are
incorrect.

This PR rewrites the test so that kernel stores both original and
broadcasted values into corresponding buffers and the former is used as
a reference for the latter and checked on host.