Replace reserve with resize for write-back buffer
In the buffer storage tests, the case using a vector iterator reserves
size in the buffer and then expects to be able to access through the
specified size. However, this is not the intention of reserve, as it
does not change the end of the vector. Instead, this should use resize.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>