Add "absolute error tolerance" accuracy mode check
Though the expected precision of math builtins are now defined by the
backend specifications, the CTS currently checks the error accuracy
based on the tolerances specified by the OpenCL C specification.
For `mix` and `smoothstep` these tolerances are defined in "absolute
error tolerance" rather than in ULP. This commit makes it possible for
the tests to switch between ULP and "absolute error tolerance" checks
based on the builtins. Likewise, it applies these to `mix` and
`smoothstep` in accordance with the OpenCL C specification.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>