Generate non-uniform group tests for group type
This commit changes most non-uniform group tests to generate separate
cases for each variant of the group types we are testing with. This was
done because the current structure will skip an entire test case if just
one of the group types are not supported by the device, even if the
others are. By splitting them, only the test case with the unsupported
group type will be skipped.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>