[NFC] Transfer host variable to sycl kernel to avoid using unsupported memory capabilities.
Test checks if memory capabilities are supported in the host code, and then
early return in the final run time for unsupported capabilities. When use
const host variable MemoryScope in the sycl kernel directly, aot
compile would fail in the compile time if MemoryScope is not supported.
So this patch transfers host variable to sycl kernel to avoid using
unsupported memory capabilities.