[range_api] Fix bitwise shift rhs out-of-range
In kernel test_range_kernels, range[2] is 64 and range_two[2] is 128.
It is undefined behavior to use either of them as rhs of bitwise shift
operator because the rhs is not smaller than the number of bits in range
value type. This PR divides rhs by 4 so that it is smaller than 64.