[NFC] Use `device::has` query directly
This patch is done with the intent to eventually drop `util/extension.h`
helper file. It probably had a reason to be there in SYCL 1.2.1 to avoid
hardcoding OpenCL extension names everywhere. However, it is very easy
to check for support for an optional feature in SYCL 2020 so there is
less value in having this verbose helper.

Note that this commit does not remove all uses of `util/extension.h` -
some of those uses are more complex than others and therefore their
removal is left to another PR.