[NFC] Drop `string_makers.h`
The only helper we had there is for printing `sycl::id` which doesn't
seem to be used anywhere, i.e. that's a dead code.

Arguably, it may be useful, but let's add it back when we will actually
need it so simplify the codebase and improve compile times (even though
ever so slightly).