[NFC] Drop `device_eval.h`
Content of the header is only used within a single test, so there is no
need for it to be located in common directory.

It seems useful at first glance, but tests for `range` are written using
different helpers which seem more generic and functional than `DEVICE_EVAL`,
so I think its fine to have the latter to be inlined into the only test
that uses it.