Speed up oneapi_memcpy2d tests by avoiding redundant setup
Due to the way the Catch2 framework interacts with `for_all_combinations`
in the SYCL-CTS, we were inadvertently re-running setup code
(allocation, fill, deallocation)
for each <T, src_ptr_type, dest_ptr_type> combination in the
oneapi_memcpy2d tests.

Add a warning to the documentation of `for_all_combinations` to highlight this
perfomance pitfall, and wrap the test code in an outer SECTION to avoid
re-executing expensive setup operations on every test case re-execution.

On my local machine with an Intel GPU, the runtime of the test
changes from 14s to 2.7s.
When running on an internal simulator it takes the test from more than
20 minutes (I killed it after that) to 9 minutes.

Disclaimer: I used a generative AI model to adjust the phrasing of the
comments, but I reviewed every line of output.